Convert PubSubTest to PubSubSpec.
